requirements.txt
# 核心依赖
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 向量处理和机器学习
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.11.0
sentence-transformers>=2.2.0

# 图数据库和向量索引
networkx>=2.6.0
faiss-cpu>=1.7.0
annoy>=1.17.0

# 数据库和存储
sqlalchemy>=1.4.0
redis>=4.0.0
pymongo>=4.0.0

# Web 框架
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# 图像处理
pillow>=8.3.0
opencv-python>=4.5.0

# 音频处理
librosa>=0.8.0
soundfile>=0.10.0

# 视频处理
moviepy>=1.0.3

# 工具库
click>=8.0.0
pyyaml>=5.4.0
python-dotenv>=0.19.0
loguru>=0.5.0

# 测试
pytest>=6.2.0
pytest-asyncio>=0.15.0
pytest-cov>=2.12.0

# 开发工具
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# LLM 客户端依赖
openai>=1.0.0
anthropic>=0.7.0
httpx>=0.24.0

# 配置和环境管理
python-dotenv>=0.19.0
pydantic-settings>=2.0.0